+++ releasenotes.md
@@ -29,6 +29,7 @@
 - Fix bug preventing optionals from being used in ranges or as indices.
 - Crash compiling for arm64 when returning 16 byte and smaller structs by value not a power of 2 #1649.
 - Enforce single module compilation for static libraries to make constructors run properly.
+- Crash when using --no-obj without compile-only. #1653
 
 ### Stdlib changes
 - Add `io::MultiReader`, `io::MultiWriter`, and `io::TeeReader` structs.